# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OCBP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/turnos/get/turns-by-date/{pavilion?}/{date?}": {
            "get": {
                "tags": [
                    "Hito"
                ],
                "summary": "get get Turns by date",
                "description": "Returns get Turns by date",
                "operationId": "get Turns by date",
                "parameters": [
                    {
                        "name": "pavilion?",
                        "in": "path",
                        "description": "Required",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date?",
                        "in": "path",
                        "description": "Required",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "X-Authorization": {
                "type": "apiKey",
                "description": "EspartaApiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "X-Authorization": []
        }
    ]
}